// src/types/components.types.ts
export type TypeResult = {
    field: string;
    label: string;
    prefix?: string;
    suffix?: string;
};

export type TypeAdResultTableSchema = {
    field: string;
    label: string;
    header: string;
    prefix?: string;
    suffix?: string;
    width?: number;
};

export enum EnumSocials {
    ig = 'IG',
    fb = 'FB',
}